requirements.txt
flask == 1.1.2
gunicorn == 20.0.4
PyYAML == 5.3.1
requests == 2.23.0
